% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_coxex.R
\name{anova.coxex}
\alias{anova.coxex}
\title{Analysis of deviance for \code{coxex} objects.}
\usage{
\method{anova}{coxex}(object, ...)
}
\arguments{
\item{object}{a \code{coxex} object.}

\item{...}{additional \code{coxex} objects.}
}
\value{
an data frame with model identifier, model formula,
and the testing results.
}
\description{
Performs an analysis of deviance table for one or more \code{coxex} class objects,
via chi-square test just like \code{\link[survival]{anova.coxph}}.
}
\details{
Please refer to the genuine method \code{\link[survival]{anova.coxph}}.
}
