% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bca}
\alias{bca}
\title{Bias-corrected and accelerated (BCA) confidence intervals}
\usage{
bca(theta, conf.level = 0.95)
}
\arguments{
\item{theta}{a numeric vector of estimates obtained in bootstrap estimates.}

\item{conf.level}{confidence interval. Defaults to 0.95.}
}
\value{
returns a numeric vector of length 2 with the lower and upper bound
of the confidence interval
}
\description{
This function uses the method proposed by DiCiccio and Efron (1996) to
compute so called bias-corrected and accelerated (BCA) confidence intervals.
The code is a modified version of the code from
https://github.com/jkropko/coxed/blob/master/R/bca.R by Jonathan Kropko and
colleagues and used by the no-more CRAN package coxed.
}
\details{
Any NAs are silently removed.
}
\references{
DiCiccio, T. J. and B. Efron. (1996). Bootstrap Confidence Intervals. \emph{Statistical Science}.
11(3): 189–212. \url{https://doi.org/10.1214/ss/1032280214}
}
\author{
Jonathan Kropko \href{mailto:jkropko@virginia.edu}{jkropko@virginia.edu} and Jeffrey J. Harden \href{mailto:jharden@nd.edu}{jharden@nd.edu}, based
on the code for the \code{\link[mediation]{mediate}} function in the \code{mediation} package
by Dustin Tingley, Teppei Yamamoto, Kentaro Hirose, Luke Keele, and Kosuke Imai.
}
