% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_oop.R
\name{brier}
\alias{brier}
\title{Build a brier class object}
\usage{
brier(times, reference, training, test = NULL)
}
\arguments{
\item{times}{a numeric vector of unique time points.}

\item{reference}{a numeric vector of Brier scores for the reference survival.}

\item{training}{a numeric vector of Brier scores for the modeled survival in
the training dataset.}

\item{test}{a numeric vector of Brier scores for the modeled survival
in the test dataset. Defaults to NULL, which means that no validation errors
are provided.}
}
\value{
an instance of the \code{brier} class with the \code{plot()} method.
The \code{brier} object is a data frame (\code{time}, \code{reference}, \code{training}
and \code{test} variables) bundling the unique times points with
their Brier scores obtained for the reference, training and test data. The
\code{brier} class inherits many of traditional data frame methods, e.g. \code{filter}
or \code{group_by} provided by the \code{dplyr} package.
}
\description{
Generates a \code{brier} class object based on a list
of unique times, Bier scores per timepoint e.g. calculated with
\code{\link[pec]{pec}} for the reference, training and test data.
}
\references{
\itemize{
\item Graf, E., Schmoor, C., Sauerbrei, W. & Schumacher, M. Assessment and
comparison of prognostic classification schemes for survival data.
Stat. Med. 18, 2529–2545 (1999).
}
}
