% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_expected_}
\alias{calc_expected_}
\title{Calculate expected values from normal distribution}
\usage{
calc_expected_(data, observed)
}
\arguments{
\item{data}{a data frame.}

\item{observed}{name of the variable storing the observed values.}
}
\value{
the input data frame with the variable \code{.expect.norm} storing the
expected normal distribution variables.
}
\description{
Computes expected normal distribution values for a sequence of
probability points (\code{\link[stats]{ppoints}})
}
