% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_calibrator.R
\name{calibrator}
\alias{calibrator}
\title{Build a calibration object.}
\usage{
calibrator(
  lp_scores,
  surv_fit,
  cox_fit,
  km_estimates,
  cox_estimates,
  strata_calibration,
  global_calibration
)
}
\arguments{
\item{lp_scores}{a data frame containing linear predictor scores.}

\item{surv_fit}{a 'surv_fit' class instance storing the Kaplan-Meier survival
estimates for the linear predictor score strata
(see: \code{\link[survminer]{surv_fit}}).}

\item{cox_fit}{a Cox proportional hazard model of the survival as a function
of the linear predictor score strata.}

\item{km_estimates}{a data frame with Kaplan-Meier survival estimates.}

\item{cox_estimates}{a data frame with cox survival estimates.}

\item{strata_calibration}{a data frame with the calibration statistics
for each linear predictor score strata.}

\item{global_calibration}{a data frame with the global
calibration statistics.}
}
\value{
an instance of 'calibrator' class.
}
\description{
Generates a calibration object,
used internally by the \code{\link{get_cox_calibration}} function.
Contains all data used in calculation of Cox model calibration with the
D'Agostino-Nam method (DOI: 10.1016/S0169-7161(03)23001-7) and visualization
of the results.
}
\references{
\itemize{
\item D’Agostino, R. B. & Nam, B. H. Evaluation of the Performance of
Survival Analysis Models: Discrimination and Calibration Measures.
Handb. Stat. 23, 1–25 (2003).
\item Royston, P. Tools for checking calibration of a Cox model in external
validation: Approach based on individual event probabilities.
Stata J. 14, 738–755 (2014).
\item Crowson, C. S. et al. Assessing calibration of prognostic risk
scores. Stat. Methods Med. Res. 25, 1692–1706 (2016).
}
}
