% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_}
\alias{count_}
\title{Count levels of a variable}
\usage{
count_(data, variable)
}
\arguments{
\item{data}{a data frame.}

\item{variable}{name of the variable of interest.}
}
\value{
a data frame with the counts.
}
\description{
Count factor variables or returns the number of complete cases
for numeric features.
}
