% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cut_quantiles}
\alias{cut_quantiles}
\title{Cut vector in n-quantile intervals.}
\usage{
cut_quantiles(x, n, labels = NULL, right = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{n}{a single numeric defining the number of quantile intervals.}

\item{labels}{an optional user-provided vector of labels.}

\item{right}{logical, indicating if the intervals should be closed on the
right (and open on the left) or vice versa.}
}
\value{
A tibble with the initial values in the order as they were provided,
the interval assignment and the user-provided labeling.
}
\description{
A numeric vector is cut in n-intervals corresponding
to quantiles (e.g: for n = 3, the vector is cut by tertiles).
A data frame is returned which contains the initial value, the interval
assignment and, optionally, the user-provided labeling.
}
\details{
the minimum and maximum values are always included in the first
and the last interval, respectively. NAs are silently removed.
}
