% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{cvCox}
\alias{cvCox}
\alias{cvCox.default}
\alias{cvCox.coxex}
\title{Cross-validation of canonical Cox proportional hazard models.}
\usage{
cvCox(data, ...)

\method{cvCox}{default}(data, formula, n_folds = 10, n_repeats = 1, ...)

\method{cvCox}{coxex}(data, n_folds = 10, n_repeats = 1, ...)
}
\arguments{
\item{data}{a data frame used to train the Cox model or
a \code{\link{coxex}} model.}

\item{...}{additional arguments passed to \code{\link[survival]{coxph}}.
Note that \code{x} and \code{y} are already used.}

\item{formula}{a formula of the Cox model; it must call a \code{Surv()} object,
i.e. contain a \code{'Surv()'} term.}

\item{n_folds}{number of cross-validation folds.}

\item{n_repeats}{number of repeats of the cross-validation procedure,
defaults to 1.}
}
\value{
a list of \code{\link{coxex}} objects - an instance
of \code{\link{mlcx}} class.
}
\description{
Cross-validation and repeated cross-validation of Cox proportional
hazard models.
}
\details{
The cross-validation procedure is essentially the same as described for
external validation with \code{\link{mlCox}}.
In brief, the training portion of each re-sample or CV fold is used to fit
a Cox model, which is used for generation of the linear predictor scores for
observations in the test portion of the re-sample. Finally, for each of
those test portions, a uni-variable Cox model is constructed which regresses
survival as a function of the linear predictor scores.
The output list consists of \code{\link{coxex}} models wrapped in
a \code{\link{mlcx}} object.
Its assumptions, fit statistics, inference, and diagnostic plots can be
accessed with \code{summary()} and \code{plot()} methods called for single elements.
Alternatively, the assumptions, fit statistics, and inference can be computed
with the dedicated \code{\link{summary.mlcx}} method.
}
