% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{formula.coxex}
\alias{formula.coxex}
\title{Modeling formula.}
\usage{
\method{formula}{coxex}(x, ...)
}
\arguments{
\item{x}{a coxex object.}

\item{...}{extra arguments, currently none.}
}
\value{
a formula
}
\description{
Accesses the modeling formula of a coxex object.
}
