% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cox_assumptions}
\alias{get_cox_assumptions}
\title{Check assumptions fo a CoxPH model.}
\usage{
get_cox_assumptions(
  cox_model,
  type.predict = "lp",
  type.residuals = "martingale",
  data = NULL,
  ...
)
}
\arguments{
\item{cox_model}{a CoxpPH model or a coxex object.}

\item{type.predict}{type of the prediction, 'lp', linear predictor score by
default. See: \code{\link[survival]{predict.coxph}} for details.}

\item{type.residuals}{type of the residuals, 'martingale' by default.
See: \code{\link[survival]{residuals.coxph}} for details.}

\item{data}{the data frame used for the model construction. Ignored,
if coxex object provided.}

\item{...}{extra arguments passed to \code{\link{get_cox_qc}}}
}
\value{
a data frame with the normality testing results and the
proportional hazard assumption testing for the model variables and the
global model.
}
\description{
Checks the normality and proportional hazard assumption for
a CoxPH model. Technically, the normality assumption is tested with
Shapiro-Wilk test, the proportionality by \code{\link[survival]{cox.zph}}.
}
\references{
\itemize{
\item Grambsch, P. M. & Therneau, T. M. Proportional Hazards Tests and
Diagnostics Based on Weighted Residuals. Biometrika 81, 515 (1994).
}
}
