% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_cox_qc_plots}
\alias{get_cox_qc_plots}
\title{Residuals plots for a CoxPH model.}
\usage{
get_cox_qc_plots(
  cox_model,
  data = NULL,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{cox_model}{a CoxpPH model or a coxex object.}

\item{data}{the data frame used for the model construction. Ignored,
if coxex object provided.}

\item{cust_theme}{a custom ggplot theme.}

\item{...}{extra arguments passed to \code{\link{get_cox_qc}}.}
}
\value{
a list of ggplot objects.
}
\description{
Generates standard residual plots for a CoxPH model: residuals
vs fitted, standardized residuals vs fitted, squared standardized residuals
vs fitted and a quantile-quantile plot of the standardized residuals.
}
