% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cox_stats}
\alias{get_cox_stats}
\title{Goodness of fit for a CoxPH model.}
\usage{
get_cox_stats(
  cox_model,
  data = NULL,
  type.predict = "lp",
  type.residuals = "martingale",
  rsq_type = c("mev", "mer", "cod"),
  ...
)
}
\arguments{
\item{cox_model}{a CoxpPH model or a coxex object.}

\item{data}{the data frame used for the model construction. Ignored,
if coxex object provided.}

\item{type.predict}{type of the prediction, 'lp', linear predictor score by
default. See: \code{\link[survival]{predict.coxph}} for details.}

\item{type.residuals}{type of the residuals, 'martingale' by default.
See: \code{\link[survival]{residuals.coxph}} for details.}

\item{rsq_type}{type of R-quared statistic, see: \code{\link[survMisc]{rsq}}.}

\item{...}{extra arguments passed to \code{\link{get_cox_qc}}.}
}
\value{
a data frame with the statistic values.
}
\description{
Calculates fit statistics for a CoxPH model: AIC (Akaike
information criterion), BIC (Bayesian information criterion), raw_rsq
(unadjusted R-squared, cod, mer or mev/default, see:
\code{\link[survMisc]{rsq}}), MAE (mean absolute error), MSE
(mean squared error), RMSE (root MSE) and C-index with 95\% confidence
intervals (normality assumption).
}
