% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_oop.R
\name{is_brier}
\alias{is_brier}
\title{Test for the brier class.}
\usage{
is_brier(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Tests if an object is an instance of the \code{brier} class.
}
