% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_calibrator.R
\name{is_calibrator}
\alias{is_calibrator}
\title{Test for the calibrator class.}
\usage{
is_calibrator(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Tests if an object is an instance of the calibrator class.
}
