% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_coxex.R
\name{is_coxex}
\alias{is_coxex}
\title{Test for the coxex class.}
\usage{
is_coxex(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Tests if an object is an instance of the coxex class.
}
