% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_mlcox.R
\name{is_mlcx}
\alias{is_mlcx}
\title{Check inheritance for \code{mlclass}.}
\usage{
is_mlcx(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Checks if an object inherits from \code{mlcx} class.
}
