% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_mlcox.R
\name{mlcx}
\alias{mlcx}
\title{Create a \code{mlcx} object.}
\usage{
mlcx(x, validation = c("cv", "external"), ...)
}
\arguments{
\item{x}{a list of \code{\link{coxex}} models.}

\item{validation}{a character specifying validation type: cross-validation
(\code{type = 'cv'}) or external validation (\code{type = 'external'}). This
information will be stored as an attribute of the object.}

\item{...}{extra arguments, currently none.}
}
\value{
a \code{mlcx} object described above with a dedicated
\code{\link{summary.mlcx}} method.
}
\description{
Creates an instance of \code{mlcx} class. Technically, it is a list of
\code{\link{coxex}} models, storing regression of survival as a function of
linear predictor scores in cross-validation folds or external data sets.
}
\details{
Intended for internal use.
}
