% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{model.frame.coxex}
\alias{model.frame.coxex}
\title{Modeling data.}
\usage{
\method{model.frame}{coxex}(formula, type = c("model_frame", "data", "surv"), ...)
}
\arguments{
\item{formula}{a coxex object.}

\item{type}{the output type: 'model_frame' returns a model frame,
'data' return the data fame used for model construction,
'surv' accesses the survival object.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame or a survival object.
}
\description{
Access to the model components and data for the coxex class.
}
