% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_calibrator.R
\name{nobs.calibrator}
\alias{nobs.calibrator}
\title{Number of observations and events.}
\usage{
\method{nobs}{calibrator}(object, ...)
}
\arguments{
\item{object}{a calibrator object.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with the number of total complete observations and
events for each linear predictor score strata.
}
\description{
Extracts the number of observations and events in
each linear predictor score strata.
}
