% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_coxex.R
\name{nobs.coxex}
\alias{nobs.coxex}
\title{Number of observations and events.}
\usage{
\method{nobs}{coxex}(object, ...)
}
\arguments{
\item{object}{a coxex object.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with the number of total complete observations and
events.
}
\description{
Extracts the number of observations and events used for
modeling.
}
