% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_oop.R
\name{plot.brier}
\alias{plot.brier}
\title{Plot a 'brier' class object.}
\usage{
\method{plot}{brier}(
  x,
  one_plot = TRUE,
  palette = c(reference = "gray60", training = "steelblue", test = "coral3"),
  linewidth = 0.5,
  show_reference = TRUE,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{x}{a \code{\link{brier}} class object.}

\item{one_plot}{logical, should Brier scores for all datasets
be presented in one plot? Defaults to TRUE.}

\item{palette}{defines colors of the Bier score curves.}

\item{linewidth}{line size.}

\item{show_reference}{logical, should the Bier score curve for the
reference be plotted? Defaults to TRUE.}

\item{cust_theme}{custom \code{ggplot} theme.}

\item{...}{extra arguments, currently none.}
}
\value{
a single \code{ggplot} graphics (if \code{one-plot} is TRUE) or a list of
\code{ggplot} plots for the Brier scores obtained for reference,
training and test data each.
}
\description{
Plots Brier scores as a function of unique time points.
}
\references{
\itemize{
\item Graf, E., Schmoor, C., Sauerbrei, W. & Schumacher, M. Assessment and
comparison of prognostic classification schemes for survival data.
Stat. Med. 18, 2529–2545 (1999).
}
}
