% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cox_fit}
\alias{plot_cox_fit}
\title{Plot outcome and expected survival.}
\usage{
plot_cox_fit(
  cox_model,
  data = NULL,
  palette = c("steelblue", "firebrick"),
  cust_theme = survminer::theme_survminer(),
  ...
)
}
\arguments{
\item{cox_model}{a CoxpPH model or a coxex object.}

\item{data}{the data frame used for the model construction. Ignored,
if coxex object provided.}

\item{palette}{color palette.}

\item{cust_theme}{custom ggplot theme.}

\item{...}{extra arguments passed to
\code{\link[survminer]{ggsurvplot_combine}}.}
}
\value{
a ggplot object.
}
\description{
Plots the actual and Cox model-predicted survival as a
Kaplan-Meier plot.
}
