% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_squares}
\alias{plot_squares}
\title{Plot squared errors between the outcome and observed survival.}
\usage{
plot_squares(
  cal_object,
  palette = c(reference = "gray60", model = "coral3"),
  show_reference = TRUE,
  error_stat = c("none", "se", "2se", "95perc"),
  ribbon_alpha = 0.25,
  line_width = 0.5,
  point_size = 2,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{cal_object}{a 'calibrator' class object.}

\item{palette}{color palette.}

\item{show_reference}{logical, should averaged squared errors be displayed
for the reference (NULL model)? Defaults to TRUE.}

\item{error_stat}{error statistic to be presented as a ribbon, defaults
to none.}

\item{ribbon_alpha}{alpha of the ribbon representing the requested error.}

\item{line_width}{width of the lines connecting the data points.}

\item{point_size}{size of the data points.}

\item{cust_theme}{custom ggplot theme.}

\item{...}{extra arguments, currently none.}
}
\value{
a list of ggplot objects.
}
\description{
Generates a series of plots of averaged squared errors as a function of
unique time points, observations, and unique values of the linear predictor
scores.
}
\details{
Intended for internal use.
Mean values are presented as points connected with lines.
Optional error region is shown as a ribbon.
}
