% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{point_plot_}
\alias{point_plot_}
\title{Generate a customized point plot.}
\usage{
point_plot_(
  data,
  x_var,
  y_var,
  x_lab = x_var,
  y_lab = y_var,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  smooth = TRUE,
  point_wjitter = 0,
  point_hjitter = 0,
  silent = TRUE,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{data}{data frame.}

\item{x_var}{name of the variable to be shown in the x axis.}

\item{y_var}{name of the variable to be shown in the y axis.}

\item{x_lab}{x axis title.}

\item{y_lab}{y axis title.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{smooth}{logical, should a trend line be displayed.}

\item{point_wjitter}{width of the point jittering, defaults to 0.}

\item{point_hjitter}{height of the point jittering, defaults to 0.}

\item{silent}{logical, display warnings?}

\item{cust_theme}{custom ggplot2 theme.}

\item{...}{extra arguments passed to geom_smooth().}
}
\value{
a ggplot graphic
}
\description{
Draws a simple point plot for model diagnostic purposes.
}
