% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_calibrator.R
\name{print.calibrator}
\alias{print.calibrator}
\title{Print a calibrator object.}
\usage{
\method{print}{calibrator}(x, ...)
}
\arguments{
\item{x}{a calibrator object.}

\item{...}{extra arguments, none specified.}
}
\value{
none, called for its side effects.
}
\description{
Prints a coxex model.
}
