% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_coxex.R
\name{print.coxex}
\alias{print.coxex}
\title{Print method for \code{coxex} objects.}
\usage{
\method{print}{coxex}(x, ...)
}
\arguments{
\item{x}{a \code{coxex} object.}

\item{...}{extra arguments, currently none.}
}
\value{
None, called for its side effects.
}
\description{
Print method for \code{coxex} class.
}
