% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sem}
\alias{sem}
\title{Standard error of the mean.}
\usage{
sem(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
a single numeric value.
}
\description{
Computes standard error of the mean (SEM) for a numeric vactor. Any NAs are
silently removed.
}
