% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_calibrator.R
\name{summary.calibrator}
\alias{summary.calibrator}
\title{Statistic summary of the D'Agostino-Nam calibration and square errors.}
\usage{
\method{summary}{calibrator}(
  object,
  type = c("global", "strata", "squares"),
  by = c("time", "observation", "lp"),
  ...
)
}
\arguments{
\item{object}{a calibrator object.}

\item{type}{type of the calibration statistic:
'global' (default) and 'strata' return the output of the D'Agostino - Nam
algorithm, while 'squares' returns a data frame with summary statistics of
square errors.}

\item{by}{type of averaging of the square errors by unique time points
('time', default), observations or unique values of the Cox model linear
predictor score ('lp').}

\item{...}{additional arguments, currently none.}
}
\value{
a data frame with the calibration statistics.
}
\description{
The function returns two types of data:
\itemize{
\item a data frame storing the numbers of events and survivors,
survival probabilities,
relative survival probability (Kaplan - Meier to Cox ratio),
relative risk (rr: Kaplan-Meier to Cox ratio) and
the D'Agostino-Nam chi-squared statistic (x2_dn) for each strata
(type = 'strata') or for the entire model (type = 'global').
\item a data frame with mean, SEM, median, and 95\% percentile
range for square errors averaged over either unique time points or
observations or unique values of the Cox model's linear predictor score.
}
}
\references{
\itemize{
\item D’Agostino, R. B. & Nam, B. H. Evaluation of the Performance of
Survival Analysis Models: Discrimination and Calibration Measures.
Handb. Stat. 23, 1–25 (2003).
\item Royston, P. Tools for checking calibration of a Cox model in external
validation: Approach based on individual event probabilities.
Stata J. 14, 738–755 (2014).
\item Crowson, C. S. et al. Assessing calibration of prognostic risk
scores. Stat. Methods Med. Res. 25, 1692–1706 (2016).
\item Graf, E., Schmoor, C., Sauerbrei, W. & Schumacher, M. Assessment and
comparison of prognostic classification schemes for survival data.
Stat. Med. 18, 2529–2545 (1999).
}
}
