% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_coxex.R
\name{summary.coxex}
\alias{summary.coxex}
\title{Summary for the coxex class.}
\usage{
\method{summary}{coxex}(object, type = c("inference", "fit", "assumptions"), ...)
}
\arguments{
\item{object}{a coxex object.}

\item{type}{statistic type, 'inference' by default.}

\item{...}{extra arguments passed to \code{\link{get_cox_estimates}},
\code{\link{get_cox_stats}} or \code{\link{get_cox_assumptions}}.}
}
\value{
a data frame with the stats specified by the type argument.
}
\description{
Generates a summary for a coxex model. This can be inference
(type = 'inference'), goodness of fit (type = 'fit') or
model assumptions (type = 'assumptions').
}
\details{
A wrapper around \code{\link{get_cox_estimates}} (inference),
\code{\link{get_cox_stats}} (goodness of fit) and
\code{\link{get_cox_assumptions}} (assumptions).
}
\references{
\itemize{
\item Therneau, T. M. & Grambsch, P. M. Modeling Survival Data: Extending
the Cox Model. (Springer Verlag, 2000).
\item Grambsch, P. M. & Therneau, T. M. Proportional Hazards Tests and
Diagnostics Based on Weighted Residuals. Biometrika 81, 515 (1994).
\item Harrell, F. E., Lee, K. L. & Mark, D. B. Multivariable prognostic
models: Issues in developing models, evaluating assumptions and adequacy,
and measuring and reducing errors. Stat. Med. 15, 361–387 (1996).
}
}
