% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{surv2df}
\alias{surv2df}
\title{Convert a survival object to a data frame with censored survival data.}
\usage{
surv2df(surv_object)
}
\arguments{
\item{surv_object}{and instance of the 'Surv' class.}
}
\value{
a data frame with the following columns:
\itemize{
\item 'time': unique time point inferred from the survival object.
\item '.observation': index of the observation in the input survival object.
\item 'status': event index as indicated in the survival object
}
}
\description{
Converts a survival object (class 'Surv') to a data frame with unique time
points and event indicators. The function accounts for censoring, i.e.
observations censored at a particular time point i do not appear at the
time point i + 1.
}
