% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/oop_coxex.R
\name{surv_brier}
\alias{surv_brier}
\alias{surv_brier.coxex}
\alias{surv_brier.coxph}
\title{Calculate Brier scores for survival models.}
\usage{
surv_brier(fit, ...)

\method{surv_brier}{coxex}(fit, splitMethod = "none", ...)

\method{surv_brier}{coxph}(fit, data, splitMethod = "none", ...)
}
\arguments{
\item{fit}{a survival model.}

\item{...}{extra arguments passed to \code{\link[pec]{pec}}.}

\item{splitMethod}{validation method used for calculation of the Bier scores
as specified by the respective \code{\link[pec]{pec}} argument.
One of 'none' (no validation), 'cvK' (K-fold cross-validation, e.g. 'cv10'),
'boot' (bootstrap), 'BootCv' (bootstrap cross-validation),
'Boot632', 'Boot632plus', 'loocv' or 'NoInf', see the upstream
function for details.}

\item{data}{the data frame used for the model construction.}
}
\value{
an object of the \code{\link{brier}} class with the \code{plot} method
and multiple methods shared with traditional data frames.
}
\description{
The Brier score for survival models is computed as a squared
distance between the predicted survival probability and and the actual
0/1-coded survival for each unique time point of the dataset.
}
\details{
See, \code{\link[pec]{pec}} for details.
The \code{surv_brier} is a S3 generic function.
}
\references{
\itemize{
\item Graf, E., Schmoor, C., Sauerbrei, W. & Schumacher, M. Assessment and
comparison of prognostic classification schemes for survival data.
Stat. Med. 18, 2529–2545 (1999).
}
}
