% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_coxex.R
\name{validate.coxex}
\alias{validate.coxex}
\title{Validate a 'coxex' model.}
\usage{
\method{validate}{coxex}(
  fit,
  method = "boot",
  B = 40,
  bw = FALSE,
  rule = "aic",
  type = "residual",
  sls = 0.05,
  aics = 0,
  force = NULL,
  estimates = TRUE,
  pr = FALSE,
  ...
)
}
\arguments{
\item{fit}{a 'coxex' model.}

\item{method}{resampling method may be "crossvalidation",
"boot" (the default), ".632", or "randomization".}

\item{B}{number of repetitions.}

\item{bw}{TRUE to do fast step-down using the \code{\link[rms]{fastbw}}
function, for both the overall model and for each repetition.
\code{\link[rms]{fastbw}} keeps parameters together that represent
the same factor.}

\item{rule}{Applies if bw = TRUE. "aic" to use Akaike's information
criterion as a stopping rule (i.e., a factor is deleted if the chi-squared
falls below twice its degrees of freedom), or "p" to use p-values.}

\item{type}{"residual" or "individual"
stopping rule is for individual factors or for the residual chi-squared
for all variables deleted}

\item{sls}{significance level for a factor to be kept in a model, or for
judging the residual chi-squared}

\item{aics}{cutoff on AIC when rule="aic"}

\item{force}{see \code{\link[rms]{fastbw}}}

\item{estimates}{see \code{\link[rms]{print.fastbw}}}

\item{pr}{TRUE to print results of each repetition}

\item{...}{extra arguments passed to \code{\link[rms]{validate.cph}}}
}
\value{
a data frame with the following variables:
\itemize{
\item \code{dataset}: dataset used for computation of the stats
\item \code{Dxy}: Somers' DXY rank correlation
\item \code{R2}: Nagelkerke R-squared
\item \code{Slope}: slope shrinkage
\item \code{D}: discrimination index D
\item \code{U}: unreliability index
\item \code{Q}: the overall quality index
\item \code{g}: g-index on the log relative hazard
\item \code{c_index}: Harrell's concordance index.
}
}
\description{
Provides an access to validation stats obtained e.g.
by cross-validation or bootstraping via \code{\link[rms]{validate}}.
}
\details{
See: \code{\link[rms]{validate.cph}}.
}
\references{
\itemize{
\item Harrell, F. E., Lee, K. L. & Mark, D. B. Multivariable prognostic
models: Issues in developing models, evaluating assumptions and adequacy,
and measuring and reducing errors. Stat. Med. 15, 361–387 (1996).
}
}
