% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cox_estimates}
\alias{get_cox_estimates}
\title{Get inference statistic for a CoxPH model.}
\usage{
get_cox_estimates(cox_model, trans_function = identity, ...)
}
\arguments{
\item{cox_model}{a CoxpPH model or a coxex object.}

\item{trans_function}{function used for transformation of the estimates and
confidence intervals, identity() by default.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with the model estimates, confidence intervals
and p values.
}
\description{
Retrieves inference statistic for a Cox proportional hazard
model. A wrapper around \code{\link[survival]{summary.coxph}}.
}
\references{
\itemize{
\item Therneau, T. M. & Grambsch, P. M. Modeling Survival Data: Extending
the Cox Model. (Springer Verlag, 2000).
}
}
