% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cox_pec}
\alias{get_cox_pec}
\title{Calculate prediction error curves and integrated Brier score (IBS).}
\usage{
get_cox_pec(cox_model, data = NULL, return_pec = FALSE, ...)
}
\arguments{
\item{cox_model}{a Cox model of the \code{coxph} or \code{coxex} class.}

\item{data}{the data frame used for the model construction. Ignored,
if coxex object provided.}

\item{return_pec}{logical, should a \code{pec} object be returned?}

\item{...}{extra arguments passed to \code{\link[pec]{pec}}.}
}
\value{
a numeric vactor with IBS values or a \code{pec} class object.
}
\description{
Computes prediction error curves and
integrated Brier score (IBS)
for a Cox model using \code{\link[pec]{pec}} and \code{\link[pec]{crps}}.
}
