% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_calibrator.R
\name{plot.calibrator}
\alias{plot.calibrator}
\title{Plot the observed and fitted survival for the linear predictor score strata.}
\usage{
\method{plot}{calibrator}(
  x,
  type = c("strata", "squares"),
  palette = NULL,
  cust_theme = survminer::theme_survminer(),
  KM_size = 0.5,
  show_cox = TRUE,
  cox_size = 0.5,
  cox_linetype = "dashed",
  color_seed = 1234,
  signif_digits = 2,
  show_reference = TRUE,
  error_stat = c("none", "se", "2se", "95perc"),
  ...
)
}
\arguments{
\item{x}{a calibrator object.}

\item{type}{type of the plot or plots:
\itemize{
\item 'strata' generates a Kaplan-Meier plot of survival in the score strata
\item 'squares' draws a series of plots of squared errors averaged over unique
time points, observations and unique values of the linear predictor score
}}

\item{palette}{a vector of color names corresponding to the strata number.}

\item{cust_theme}{custom ggplot theme.}

\item{KM_size}{size of the Kaplan-Meier plot line, ignored
if \code{type = 'squares'}.}

\item{show_cox}{logical, should the Cox model-predicted survival for
the strata be plotted? Ignored if \code{type = 'squares'}.}

\item{cox_size}{size of the Cox survival plot line.
Ignored if \code{show_cox = FALSE} or \code{type = 'squares'}.}

\item{cox_linetype}{type of the Cox survival plot line, dashed by default.
Ignored if \code{show_cox = FALSE} or \code{type = 'squares'}.}

\item{color_seed}{seed for the random color generator, ignored
if palette provided.}

\item{signif_digits}{significant digits for rounding the calibration
statistics displayed in the plot caption. Ignored if \code{type = 'squares'}.}

\item{show_reference}{logical, should averaged squared errors be displayed
for the reference (NULL model)? Defaults to TRUE. Used only if
\code{type = 'squares'}.}

\item{error_stat}{error statistic to be presented as a ribbon, defaults
to none. Used only if \code{type = 'squares'}.}

\item{...}{extra arguments passed to \code{\link[survminer]{ggsurvplot}}
(if \code{type = 'strata'}) or to \code{\link{plot_squares}}.}
}
\value{
a ggplot object or a list og ggplot objects.
}
\description{
Draws two types of plots:
\itemize{
\item a Kaplan-Meier plots of the observed survival (solid line)
and the Cox model-predicted survival in each linear predictor score strata.
The global calibration measures are presented in plot caption. The legend
contains the number of observations assigned to each linear predictor score
strata. The total numbers of observations and events are shown in the plot
tag.
\item a list of plots of square errors averaged over unique time points,
observations and unique values of the linear predictor score
}
}
\details{
The Kaplan-Meier plot is generated with
\code{\link[survminer]{ggsurvplot}}.
}
