% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_calibrator.R
\name{plot.calibrator}
\alias{plot.calibrator}
\title{Plot the observed and fitted survival for the linear predictor score strata.}
\usage{
\method{plot}{calibrator}(
  x,
  palette = NULL,
  cust_theme = survminer::theme_survminer(),
  KM_size = 0.5,
  show_cox = TRUE,
  cox_size = 0.5,
  cox_linetype = "dashed",
  color_seed = 1234,
  signif_digits = 2,
  ...
)
}
\arguments{
\item{x}{a calibrator object.}

\item{palette}{a vector of color names corresponding to the strata number.}

\item{cust_theme}{custom ggplot theme.}

\item{KM_size}{size of the Kaplan-Meier plot line.}

\item{show_cox}{logical, should the Cox model-predicted survival for
the strata be plotted?}

\item{cox_size}{size of the Cox survival plot line.
Ignored if \code{show_cox} is set to FALSE.}

\item{cox_linetype}{type of the Cox survival plot line, dashed by default.
Ignored if \code{show_cox} is set to FALSE.}

\item{color_seed}{seed for the random color generator, ignored
if palette provided.}

\item{signif_digits}{significant digits for rounding the calibration
statistics displayed in the plot caption.}

\item{...}{extra arguments passed to \code{\link[survminer]{ggsurvplot}}.}
}
\value{
a ggplot object.
}
\description{
Draws a Kaplan-Meier plots of the observed survival (solid line)
and the Cox model-predicted survival in each linear predictor score strata.
The global calibration measures are presented in plot caption. The legend
contains the number of observations assigned to each linear predictor score
strata.
The total numbers of observations and events are shown in the plot tag.
}
\details{
The Kaplan-Meier plot is generated with
\code{\link[survminer]{ggsurvplot}}.
}
