% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_coxex.R
\name{predict.coxex}
\alias{predict.coxex}
\title{Prediction for the coxex class.}
\usage{
\method{predict}{coxex}(object, ...)
}
\arguments{
\item{object}{a coxex object.}

\item{...}{additional arguments passed
to \code{\link[survival]{predict.coxph}}.}
}
\value{
a vector or matrix of predictions, or a list containing the
predictions.
}
\description{
Predicts survival, linear predictors etc. for a coxex model.
Technically, a wrapper around \code{\link[survival]{predict.coxph}}.
}
