% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_coxex.R
\name{print.coxex}
\alias{print.coxex}
\title{Print a coxex model.}
\usage{
\method{print}{coxex}(x, ...)
}
\arguments{
\item{x}{a coxex object.}

\item{...}{extra arguments, none specified.}
}
\value{
none, called for its side effects.
}
\description{
Prints a coxex model.
}
