% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_calibrator.R
\name{summary.calibrator}
\alias{summary.calibrator}
\title{Statistic summary of the D'Agostino-Nam calibration.}
\usage{
\method{summary}{calibrator}(object, type = c("global", "strata"), ...)
}
\arguments{
\item{object}{a calibrator object.}

\item{type}{type of the calibration statistic: 'global' (default) or 'strata'.}

\item{...}{additional arguments, currently none.}
}
\value{
a data frame with the calibration statistics.
}
\description{
Returns a a data frame storing the numbers of events
and survivors, survival probabilities,
relative survival probability (Kaplan - Meier to Cox ratio),
relative risk (rr: Kaplan-Meier to Cox ratio) and
the D'Agostino-Nam chi-squared statistic (x2_dn) for each strata
(type = 'strata') or for the entire model (type = 'global').
}
