% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{summary.coxex}
\alias{summary.coxex}
\title{Summary for the coxex class.}
\usage{
\method{summary}{coxex}(object, type = c("inference", "fit", "assumptions"), ...)
}
\arguments{
\item{object}{a coxex object.}

\item{type}{statistic type, 'inference' by default.}

\item{...}{extra arguments passed to \code{\link{get_cox_estimates}},
\code{\link{get_cox_stats}} or \code{\link{get_cox_assumptions}}.}
}
\value{
a data frame with the stats specified by the type argument.
}
\description{
Generates a summary for a coxex model. This can be inference
(type = 'inference'), goodness of fit (type = 'fit') or
model assumptions (type = 'assumptions').
}
\details{
A wrapper around \code{\link{get_cox_estimates}} (inference),
\code{\link{get_cox_stats}} (goodness of fit) and
\code{\link{get_cox_assumptions}} (assumptions).
}
